Fixed bug in test search500.
When searching over the latest 500 annotations in hypothesis there is a possibility that while searching somebody has created a new annotation and that one is moves the rest in the queue to a later position, so when retrieving the next 200 in the next query there are some repeated annotations in the list. That's why the searchAnnotations not always returns the specified limit. The solution to reduce this problem is to searchAfter a certain period of time.